requirements.txt
flask
pdfminer.six
requests
chromadb